add .adt pretty-printer in format_adt.py
This makes it easier to inspect .adt files manually without the burden
of fully parsing them.

The formatter works by exploiting the almost-Python syntax of the ADT.
It tweaks it into valid Python, then parses and formats it using
Python's ast.literal_eval and pprint.pprint modules. After this, it
reverses the transformation to produce something very much like BAP.