Return Some[] from irrefutable unapply methods
this changes the return type of irrefutable unapply methods
to be Some[...] instead of Option[...], aiding the compiler's
deduction of pattern-match exhaustivity.

this avoids confusing warnings where there is a case

    case LocalVar(_, _, _) => ...

but the compiler still says

    "it would fail on pattern case: _: ir.LocalVar",

presumably accounting for the situation where the unapply method
returns None.

with this change, we are able to remove one "unreachable code"
assertion in GTIRBToIR.scala, as the compiler correctly deduces
it is unreachable.