derive for Command by making a fake IR
also need to remove Excl <: T bound from derivers. idk why

also, variance is kinda tricky. at the moment, ToScala is contravariant
(because it is a function from T => String). however, the result of
ToScala needs to be a scala string which evaluates into the original
type. so, i think it should actually be invariant

e.g. ToScala[Map] is invoked for SortedMap values. this will produce
non-compiling code if a constructor requires a SortedMap.