move ToScalaLines[MemorySection]
we need to neatly implement this dependency hierarchy of functions
(where ToScala of each level depends on ToScala of smaller levels)
(in the program, procedure, block hierarchy).

i plan to do this by making each function take an argument which
contains every function in the group.

or, we can try to make it work with inheritance.