output initial memory in program.toScala (#338)
* output initial memory in program.toScala

* ToScalaDeriving: handle SingletonProxy as well, add comment

not handling SingletonProxy was leading to producing "None()"
instead of "None". not clear why this only appeared after
the separate None/Some ToScala instances.

* ToScala[MemorySection]: refactor, omit memory if empty

the output is slightly changed. bytes are always padded to 2 hex
digits, and the readonly/region fields are on their own lines
with parameter names specified.

* ToScalaTest: add test for prog toscala with memory

* move ToScalaLines[MemorySection]

we need to neatly implement this dependency hierarchy of functions
(where ToScala of each level depends on ToScala of smaller levels)
(in the program, procedure, block hierarchy).

i plan to do this by making each function take an argument which
contains every function in the group.

or, we can try to make it work with inheritance.

* first pass using trait inheritance for overriding behaviour

i'm so silly. inheritance naturally gives us overriding of methods
and access to the original implementation of methods (via super).
this should make things a bit cleaner

* TODO: think about where to insert cut point for initial memory

idk do we just add a dummy implementation in BasilIRToScala??
i don't want to have to repeat the rest of the programToScala

* code is working with multiple trait inheritance

TODO: comment comment comment

* doc comments

* review comments

---------

Co-authored-by: rina <k@rina.fyi>