use .to(ArraySeq) instead of .toArray, fixing some deprecations (#365)
The .toArray method converts an iterable to an Array. However,
this is a minimal Java-esque array and does not support any of Scala's
interfaces. Where an interface like Iterable or Seq was needed, this
would require a coercion from Array to ArraySeq, which the compiler
tells us would incur an additional copy.

This fixes the deprecation warning by using converting directly to
ArraySeq, avoiding the implicit coercion and additional copy.