Tagging of test suites (#350)
In the Github actions, test cases are grouped into a number of coarse categories (unit tests, standard system tests, and analysis system tests). This PR aims to make these groups more concrete by defining them within the test files, instead of in the Github actions workflow. This will also allow you to run a particular group on your local machine while excluding the slower-running tests.

These tags are defined as annotations within the test_util.tags package, and an annotation can be attached to any test suite class:

@test_util.tags.UnitTest
class IRTest extends AnyFunSuite { ... }

When running tests, one can include or exclude a particular tag by using the -n or -l flags of the scalatest Runner. With SBT, these arguments can be affixed to the testOnly command:

sbt "testOnly * -- -n test_util.tags.UnitTest"  

With mill, it is more complicated. Adding -n or -l to ./mill test.testOnly will not work :( (all test suites will still be run). I think there is a problem with how Mill invokes the scalatest runner. The workaround is to manually invoke the runner:

./mill test.runMain org.scalatest.tools.Runner -o -R out/test/compile.dest/classes -n test_util.tags.UnitTest

After the -R .../classes, the usual test.testOnly arguments can be added. Note that the -o flag is important - without it, scalatest will try (and likely fail) to open a GUI. The -R argument should point to the output directory containing compiled class files.

The Github CI jobs are changed to use the ./mill test.runMain command. Also, there is a new CI job which reports the test suites which do not have any tag. It is my hope that every test case will eventually have at least one tag - I will need some help with this. This will make sure that every test written is being exercised by the CI. Currently, it is too easy for tests to fall through the craks if they are not run regularly.

This PR is part of an effort to improve the testing in Basil and force us to be more honest about the tests. This means that we should have a clear understanding of what each test is testing, whether it currently passes and, if not, why it is currently not working. I will point out these scalatest features to help this goal:

    pendingUntilFixed should be used to mark a block of code (typically containing an assertion) as one which is currently expected to fail. This should be used to record test cases which fail due to not-yet-implemented features or known bugs. If a change is made and the code no longer fails, this will cause the test to fail until the pendingUntilFixed is removed. It should be used within a test() { ... } block like so:

    test("1 == 2 sometime soon?") {
      assert(1 == 1, "obviously")
      pendingUntilFixed {
        assert(1 == 2, "todo")
      }
    }

    Tests can be marked as ignored by replacing test() with ignore(). An entire suite can be marked as ignored with the @org.scalatest.Ignore annotation.

    We also add a trait trait TestCustomisation to help with customising particular dynamically-generated tests based on their test case name (which includes file path and compiler/lifter options).

    Individual test cases can be marked as retry, disabled, or expect failure. This modified the behaviour of the test case and prints helpful output when running the test. For example:

- correct/malloc_with_local3/clang:BAP (pending)
  + NOTE: Test case is customised with: "ExpectFailure(previous failure was: Expected verification success, but got failure. Failing assertion is: assert (load37_1 == R30_in))"
 
  + Failing assertion ./src/test/correct/malloc_with_local3/clang/malloc_with_local3_bap.bpl:264 
  + 261 |     load36_1, Gamma_load36_1 := memory_load64_le(mem_9, bvadd64(R31_in, 18446744073709551600bv64)), (gamma_load64(Gamma_mem_9, bvadd64(R31_in, 18446744073709551600bv64)) || L(bvadd64(R31_in, 18446744073709551600bv64)));
    262 |     call rely();
    263 |     load37_1, Gamma_load37_1 := memory_load64_le(mem_10, bvadd64(R31_in, 18446744073709551608bv64)), (gamma_load64(Gamma_mem_10, bvadd64(R31_in, 18446744073709551608bv64)) || L(bvadd64(R31_in, 18446744073709551608bv64)));
 >  264 |     assert (load37_1 == R30_in); //is returning to caller-set R30
    265 |     goto printCharValue_2260_basil_return;
    266 |   printCharValue_2260_basil_return:
    267 |     assume {:captureState "printCharValue_2260_basil_return"} true; 

- analysis_differential:malloc_with_local/gcc_O2:GTIRB (pending)
  + NOTE: Test case is customised with: "ExpectFailure(needs printf_chk)"
 
  + STDOUT: "" 

Note: The tag annotations must be defined in Java (apparently) to make them visible to the test runner at runtime.


* tagging up to unit tests

* add standardsystemtests and comment for annotatetestcase

* ci: check for test tagging

* ci: use tags to run tests

* scalafmt

* ci: update message

* use cancel() instead of assume(false)

* DifferentialTest marking. added TestCustomisation mixin

* tag remaining tests

* add script. add some missing tags

* ci: add durations to test runner output

* SATTest: remove smt.timeout option (it is set by command line)

* disable stack_pointer cases in memory region tests

* extra spec tests

* indirect call tests

* tweak customised test messages, fix SimplifyMemorySystemTests

* fix testcustomisation case oopsie, use self-type instead of AnyFunSuite

* print current outcome of xfail tests

* review: split ExpectFailure into TempFailure and NotImplemented

* move TestCustomisation into its own file

* tests: update indirect call test modes after merge

* scalafmt

* add script to execute scalatest runner

this saves having to use the ./mill test.runMain ... ... command

* IndirectCallTests: add remarks from review

* add Slow tag, make check-test-tagging stricter

* tag IntervalDSATest

* IndirectCallTests: move to Analysis and un-mark some not implemented

* add some helpful commands to scalatest.sh

* docs: add testing.md to describe some testing infrastructure

* IntervalDSATest: mark as broken

* IntervalDSATest: unbreak one test

* IntervalDSATest: unbreak