tests: thread-safe test suites (#353)
This makes steps towards making test suites thread-safe. That is, multiple suites can be executed in parallel, but test cases within the same suite are still run sequentially.

* add locks around systemtests

* DataStructureAnalysis: avoid global NodeCounter

this replaces the global singleton NodeCounter with
a new instance per DSA execution. this allows multiple
DSA runs to run in parallel.

the NodeCounter is constructed by the DataStructureAnalysis class,
then passed through an implicit "using" parameter to the classes
which require it.

* remove reset method from NodeCounter

* refactor to LockManager class

* lock around testResults

* scalafmt

* greatly simplify Lockmanager by not using RWLock

LockManager now uses a single lock around the map and this is shared
for reads and writes. this will increase lock contention. however,
the locked duration will be extremely small since it only wraps
a single hashmap operation. this should have no discernible performance
impact.

* Apply suggestions from code review

* SystemTests: shuffle tests

honestly, i think this might be annoying, especially if you try
to run the tests locally to reproduce an issue and the order keeps
changing.

* SystemTests: insert suite name into output file names

this should avoid clashes where test suites try to write
to the same output file.

SystemTestsBAP and SystemTestsGTIRB are overriden to have
no suffix, to maintain compatibility with the update-expected
sbt/mill jobs.

* Revert "SystemTests: shuffle tests"

This reverts commit f4b3a53a57ece5914a7c49f79ad5ec146a50bff2.

* ci: run 3 suites in parallel

-T is used to re-sort the test cases into the original order
before printing them in batches for each suite to avoid interleaving
individual cases from multiple suites running in parallel.

* RegionTimer: use atomic long instead of instance vars

* ExtraSpecTests: retry on failure