micro-opt: cache hashcode of immutable case classes
We cache the hashCode() result of case classes the first time they are
computed. Given the prevalence of hashmaps, this is a very frequent
operation and hashCode of strings and bigints is slow (enough) that
recomputing it too often is detrimental to performance.

IMPORTANT: this caching must /not/ be applied to case classes with
           mutable fields! doing so would lead to inconsistencies between
           equals and hashcode.