micro-opt: reduce hashmap lookups in CopyProp
We make some very minor changes to the simp CopyProp to reduce the number
of hashmap searches needed. This is a hotspot and each individual
`hashmap(l)` performs a search of the hashmap. It is an unnecessary
performance penalty to do hashmap.contains() immediately followed by
multiple hashmap(key) operations.