micro-optimisations: cached hashcode and reduce hashmap lookups (#381)
* micro-opt: cache hashcode of immutable case classes

We cache the hashCode() result of case classes the first time they are
computed. Given the prevalence of hashmaps, this is a very frequent
operation and hashCode of strings and bigints is slow (enough) that
recomputing it too often is detrimental to performance.

IMPORTANT: this caching must /not/ be applied to case classes with
           mutable fields! doing so would lead to inconsistencies between
           equals and hashcode.

* micro-opt: reduce hashmap lookups in CopyProp

We make some very minor changes to the simp CopyProp to reduce the number
of hashmap searches needed. This is a hotspot and each individual
`hashmap(l)` performs a search of the hashmap. It is an unnecessary
performance penalty to do hashmap.contains() immediately followed by
multiple hashmap(key) operations.

* micro-opt: no need to .toSeq in ILLexicalIterator.next()

pushAll is happy to take an Iterator as argument and we can append
to an Iterator using ++.