add: dominator tree computation (#385)
This implements a simple function to compute dominators of a block.
A dominator of a block `b` is a block `b'` such that every path
from the entry (of a procedure) to `b` first passes through `b'`.

To determine whether two arbitrary blocks dominate each other,
you should first compute the dominator tree. Then, pass this to
`Dominator.dominates` to obtain a binary predicate on blocks.

This follows the algorithm in https://hdl.handle.net/1911/96345.
Although this runs in quadratic time, the paper notes that it is
simpler and empirically faster than asymptotically faster algorithms
for realistic sizes of programs.