tests: redirect test stdout/err to info() messages  (#372)
* tests: redirect test stdout/err to info() messages

This helps organise the test output, especially when running in
parallel. With this change, stdout/stderr are redirected through
scalatest's info() method, allowing the output to be associated and
printed with the name of the test case.

* tests: mix in test_util.CaptureOutput to redirect stdout/stderr

* scalafmt + comments

* ci: fix multiline class defns in check-test-tagging.sh

this modifies the regex which looks for "class NAME " to additionally
match an end of line after NAME.

* logging: restore setOutput, properly propagate by-name param

Each PrintStream param needs to be passed by-name to allow for the
original expression to be propagated through all the helper methods and
overloads.

By changing the `output` field to a parameterless method, this also allows
us to bring back the setOutput function.