gtirb: add a word in front of bytestring uuids, for readability
this uses the first two bytes of the GTIRB block UUID to choose one of
7776 words, then prefixes this in front of the base64 string. this
results in block names like:

```
main_1812__2__sacred_UnO2T_q8SO22cZ~w_xJwoA
main_1812__1__deceptive_22DSnEAzTL6vU9qI4utrZA
main_1812__0__penpal_z7dsAZ41QYe9rvbDt1JoIQ
```

i think this would improve "at-a-glance" readability of the IR and Boogie
output, since words are more recognisable than base64 strings. this can
also make blocks easier to search. you can search just the word and
you have a good chance of finding what you want - this can even be done
by typing it manually instead of copy/pasting.

the main disadvantage of this is the relative small size of the word
list. with only 7776 words, there is a 47% probability of collision with 100
words (assuming the words are uniformly chosen, which is reasonable
given the randomly-generated uuids). we could use two words and this
increases the blocks needed for a 48% collision probability to 9000.
however, two words is slightly more unwieldly.

the words are also only added onto block UUIDs. where more Basil IR blocks
are derived from the same GTIRB blocks, these are usually suffixed with
a counter. maybe we can add a word to the counter as well.

i am not too attached to this PR so you can take it or leave it.

here is an excerpt of a basil ir program showing what it looks like with
multiple blocks:
```
proc main_1812() -> ()
{
  name = "main";
  address = 0x714;
  entry_block = "main_1812__0__penpal_z7dsAZ41QYe9rvbDt1JoIQ";
  blocks = [
    block main_1812__0__penpal_z7dsAZ41QYe9rvbDt1JoIQ {address = 0x714} [
      goto(main_1812__0__penpal_z7dsAZ41QYe9rvbDt1JoIQ$__0, main_1812__0__penpal_z7dsAZ41QYe9rvbDt1JoIQ$__1)
    ];
    block main_1812__0__penpal_z7dsAZ41QYe9rvbDt1JoIQ$__1 [
      goto(main_1812__0__penpal_z7dsAZ41QYe9rvbDt1JoIQ_goto_main_1812__3__capable_AzuUI8EJR~~zZ4LRu_atQA, main_1812__0__penpal_z7dsAZ41QYe9rvbDt1JoIQ_goto_main_1812__1__deceptive_22DSnEAzTL6vU9qI4utrZA)
    ];
    block main_1812__0__penpal_z7dsAZ41QYe9rvbDt1JoIQ$__0 [
      goto(main_1812__0__penpal_z7dsAZ41QYe9rvbDt1JoIQ_goto_main_1812__3__capable_AzuUI8EJR~~zZ4LRu_atQA, main_1812__0__penpal_z7dsAZ41QYe9rvbDt1JoIQ_goto_main_1812__1__deceptive_22DSnEAzTL6vU9qI4utrZA)
    ];
    block main_1812__0__penpal_z7dsAZ41QYe9rvbDt1JoIQ_goto_main_1812__1__deceptive_22DSnEAzTL6vU9qI4utrZA [
      assume boolnot(bveq(extract(1, 0, R8:bv64), 0x1:bv1));
      goto(main_1812__1__deceptive_22DSnEAzTL6vU9qI4utrZA)
    ];
    block main_1812__1__deceptive_22DSnEAzTL6vU9qI4utrZA {address = 0x738} [
      goto(main_1812__2__sacred_UnO2T_q8SO22cZ~w_xJwoA)
    ];
    block main_1812__2__sacred_UnO2T_q8SO22cZ~w_xJwoA {address = 0x73c} [
      goto(main_1812__4__agonizing_hiOAUbGYTmCyjtnuNPouiQ)
    ];
    block main_1812__0__penpal_z7dsAZ41QYe9rvbDt1JoIQ_goto_main_1812__3__capable_AzuUI8EJR~~zZ4LRu_atQA [
      goto(main_1812__3__capable_AzuUI8EJR~~zZ4LRu_atQA)
    ];
    block main_1812__3__capable_AzuUI8EJR~~zZ4LRu_atQA {address = 0x748} [
      goto(main_1812__4__agonizing_hiOAUbGYTmCyjtnuNPouiQ)
    ];
    block main_1812__4__agonizing_hiOAUbGYTmCyjtnuNPouiQ {address = 0x754} [
      goto(main_1812_basil_return)
    ];
    block main_1812_basil_return [
      return ()
    ]
  ];
};
```

the source of the words is EFF's long wordlist: https://www.eff.org/dice

to explore the collision probabilities, you can use this calculator:
https://kevingal.com/apps/collision.html