fix GenericLogger.writeToFile bug which produced empty file (#406)
* fix GenericLogger.writeToFile bug which produced empty file

I'm very sorry about this one.
In my changes to GenericLogger to implement the redirection of
stdout/stderr, I changed the PrintStream parameter to be a by-name
(`=>`) parameter.

This means it is evaluated every time the parameter
is used. This was needed to allow for the possibility that Console.out,
for example, would be redirected to a new stream.

However, in this instance, it causes problems when the by-name
parameter is re-evaluated on every call. Here, the `PrintStream(file)`
parameter being re-evaluated means that new PrintStreams are being
constructed on the same file, leading to conflicts when flushing and
closing, and extremely surprising behaviour.

This was a nasty one, I'm sorry.

* avoid by-name parameters in GenericLogger methods

by-name parameters are replaced with separate overloads
which take either a by-value parameter or a lambda function
to produce the PrintStream.