tests: support scaled timeouts to fix CI timeouts  (#386)
* tests: support scaled timeouts to fix CI timeouts

This one correct/secret_write/clang_pic:BAP test has been failing on and
off since at least https://github.com/UQ-PAC/BASIL/pull/381. At the
moment, in CI, it runs in about 10 seconds which is just at the border
of the standard Boogie timeout.

This PR changes the Boogie timeouts to use scalatest scaled time spans,
allowing timeouts to be scaled by a multiplicative factor through the
command line.

We use this to double the CI's boogie timeout to at least 20 seconds.

* ci: double AnalysisSystemTest timeouts

Hopefully, this should reduce flakiness of correct/secret_write/clang_pic:BAP.
The increase in timeouts should not negatively impact performance, as it
is only a timeout.

* add some missing CaptureOutput mixins