use reaching defs to propagate definition only when dsa invariant is maintained by the code motion
my hope is that all definitions of dependent variables being reachable
from the use introduced by the propagation is sufficient to show the
variable isn't 'clobbered'.

This really feels like a hack, reaching-definitions is fairly
inefficient and we need to cull the domain by the
live variables for it to terminate on cntlm in a reasonable time.
For this to work for flag propagation we introduce the depenencies of the
propagated expression into the live vars domain used to cull the
reaching defs set, for every guard.