Support for rely-guarantee generation (#400)
introduces "interference domains" for abstracting interferences between threads, and provides a function for generating sound and compatible rely-guarantee conditions for concurrent programs using these domains.

This work is still in development. For the moment, we don't work on top of memory-region analysis, we ignore memory reads and writes, and we don't distinguish between global and local variables. We also don't provide a good way to define each thread (during testing, we just assume a thread is some given Procedure), and there are still a number of questions about the implementation of the interference domains that need to be resolved before deployment. The contribution of this PR is the core logic of the rely-guarantee generator and of a number of interference domains.

---------

Co-authored-by: bpaul <bpaul@bpaul.xyz>
Co-authored-by: alistair <alistair.michael@uq.edu.au>