move RemovePCStatements into load() ;-;
this fixes the DifferentialAnalysisTest seeing program counters
in its loaded IR, but it is maybe undesirable to do a transformation
in the load() method.

regardless, the current approach (of keep PC, then remove if unwanted)
seems to necessitate this, as later parts of the runutils pipeline
make assumptions that PC wont be present and fail in unexpected ways
otherwise.