track PC and assert at beginning of blocks  (#387)
* feat: maintain program counter and assert at beginning of blocks

this makes a fairly significant change to the GTIRB loader by inserting
PC increment statements after instructions and checking the PC at the
entry of blocks. this is done by changing the loader to maintain PC and
branchtaken assignments, and inserting a default increment if there is
no branch.

* apportion TempIf statements into successor blocks, then remove TempIf

this fixes the problemetic cross pattern we saw, where there were extra
blocks inserted before the original successor blocks.

* add RemoveUnreachableBlocks pass to clean up padding blocks



* move RemovePCStatements into load()

this fixes the DifferentialAnalysisTest seeing program counters
in its loaded IR, but it is maybe undesirable to do a transformation
in the load() method.

regardless, the current approach (of keep PC, then remove if unwanted)
seems to necessitate this, as later parts of the runutils pipeline
make assumptions that PC wont be present and fail in unexpected ways
otherwise.

* add back RemoveUnreachableBlocks to loadAndTranslate

* update expected. some differences in memory sizes and function inlining?

* change --pc argument to an enum

* implement Assert feature, use requires/ensures instead of assume

* use "pc-tracking" label

* scalafmt

* add PCTrackingTest, using a test case which has a runtime-dependent branch

* inject requires/ensures through Specification logic

to hopefully make sure it's picked up by the parameter form transform.
despite this, it looks like it still is not working.

---------

Co-authored-by: am <alistair.michael@uq.edu.au>