migrate json to uPickle and small refactor
spray-json is functionally unmaintained, not having a change
since 4 years ago. this migrates the JSON parsing to use Li Haoyi's
uPickle library. currently, the only JSON in the project is within
the GTIRB semantics loader (but maybe this will increase in future).

in addition to seeing recent updates, uPickle presents a more concise
and composable interface. for example, it can parse directly into
Scala datatypes of Map and List without needing an extra conversion
step.

i also take the opportunity to add a new functional programming
function, `sequence`. this converts a list of option to an option of
list - returning Some if all the options are Some, and returning
None if any of the options were None. previously, this logic was
handled by a `.exists(_.isEmpty)` then a `.map(_.get)`.