parallelise tests using testForkGrouping and post-processing the output (#429)
* add new scalatest.sh to rearrange `./mill test` output

When testForkGrouping is used, the output would usually be
interspersed due to the multiple concurrent processes (as opposed to threads).
This script will re-group each process's output, making use of the "[001]" prefix which Mill
inserts on each line.

* remove AnalysisSystemTest2,3,4 and use new scalatest.sh

* set TERM before invoking tput

* disable parallel tests unless BASIL_TEST_PARALLEL=true

this preserves the traditional behaviour for local test execution
and on the windows runner. this flag is only set by the scalatest.sh
script which knows how to handle the parallel test output.

* try in 3 groups

* Update scalatest.sh

* Update build.mill

* Update build.mill