ci: use scalatest runner directly (again)
this brings back the old scalatest.sh which directly
invokes scalatest's runner. now that there's a case
fallback if MILL_WORKSPACE_ROOT is unset, this seems
to work find and leads to the best behaviour.

tests are run all in one thread, allowing for minimal
JVM overhead and well-sorted output.