Make --relf optional (#448)
Make --relf optional for tasks not requiring symbols. The symbol table
is only really required for the spec input and DSA heuristics and the
relf parser tends to be by far the most fragile part of the frontend.

When relf is not provided look for `--main-procedure-name` (default
'main') in the procedure list to set as Program.mainProcedure. This is
previously all we do to detect the entrypoint, except with the relf
symbol table to find a procedure address to search for.

I think eventually we should be able to retrieve all the same
information from the gtirb `Module.symbols` but that's a bigger task.

---------

Co-authored-by: rina <k@rina.fyi>