ci: use scalatest runner directly (again) (#447)
this brings back the old scalatest.sh which directly invokes scalatest's
runner. now that there's a case fallback if MILL_WORKSPACE_ROOT is
unset, this seems to work fine and leads to the best behaviour:

tests are run all in one thread, allowing for the best runtime
performance and correctly-sorted output.

for reference:
- commit before mill 12.10 and scalatest.sh removal
https://github.com/UQ-PAC/BASIL/commit/55587f08f4f90ecd35e86c14c8729669f67ca0b3
- my unfortunate suggestion to use `./mill test` directly
https://github.com/UQ-PAC/BASIL/pull/421#issuecomment-2861915997
- hacky band-aid by re-sorting output. this worked but it was slow.
https://github.com/UQ-PAC/BASIL/pull/429

i'm sorry for the trouble lol