add HTML and PDF output for the BNFC grammar (#465)
This processes the BNFC grammar through its Latex output and converts
that to HTML using lwarp. This is added to the Basil docs website.

This makes the docs building process more complicated, since it needs
texlive. The Github Actions are split up to factor out the apt
installations. The test workflow now invokes the docs.yml workflow.


![image](https://github.com/user-attachments/assets/c86ec24a-b76c-4754-929f-ebc91b014031)