Delete old java-style visitor (#466)
Ports the last couple of transforms to the new CILVisitor and deletes
the old visitor.

CILVisitor is generally more idiomatic functional code and more powerful
than the existing visitor (and requries less cruft to be distributed
about the AST).

This should produce identical behaviour to before. 

Since translating.ILToIL() is not really used and is replaced by
PrettyPrinter it just deletes this instead of porting it.