BIG! abolish ByteString and use Base64-in-String throughout
the toString of ByteString is, i assume, some java-specific format
with ascii + escape codes. pervasively using ByteString causes huge
headaches when you're trying to match uuids from basil with uuids
from the GTIRB tools which all use base64-encoded bytes.

this pr changes the gtirb loader to convert bytestrings to base64
as soon as possible, then uses the base64 in a string for all other
uses. hopefully, this behaves identically to the previous code.

it might be worth discussing whether we want a special case class
or opaque type to store the base64 uuid, to avoid possible confusion
with other string types.