add `--version` using git tags only (#483)
Adds a `--version` argument to Basil which records the output of `"git"
"describe" "--tags" "--dirty"` into the compiled binary. If Git is not
present, a `COMMIT` environment variable must be present to specify this
information.

Also add a short docs page about this.

Does not add any VERSION file into the repository.

---------

Co-authored-by: am <alistair.michael@uq.edu.au>