attrib parsing moved. but parsing Prog Decls in main visitor is ass,
because they all need to return an EventuallyProcedure? but they
have entry block and attribs, which should go into a ProgSpec,
but they also have rely/guarantee which needs to be expression-aware.
that is, after the arly phase.

maybe just split this across the two visitors - do most things early,
but the rely/guar in mainvisitor