BNFC grammar and visitor refactor (#488)
update grammar rules to be more consistent, with cases prefixed by their
label. in the scala code, this makes it much easier to know which rule
each visit method corresponds to.

refactor visitors to bring back mutability and factor things out.

now, in the early phase:
- earlybnfcvisitor
- declsexprvisitor is an inner class in the early visitor. this
constructs an exprbnfcvisitor to visit the top-level expression
predicates.

in the main phase (indentation should be taken to mean the upper one
"uses" the lower one).
- mainbnfcvisitor - visits programs. also changed to return DSL again,
with a function in ParseBasilIL which builds the IRContext.
- blockbnfcvisitor - needs procedure name and decls, visits statements
and blocks
    - exprbnfcvisitor - needs decls, visits expressions
      - typesbnfcvisitor, attriblistvisitor, etc.

moves all declaration and spec information into the Declarations case
class. also makes it store only formal ins/outs instead of an
EventuallyProcedure, to make it clearer what it does.