Java SMT interface (#501)
Creates a wrapper around
[java-smt](https://github.com/sosy-lab/java-smt/) which is an interface
to numerous SMT solvers. At the moment just Z3 is being used, but it
won't be much effort to install more of them.

At the time of writing this Z3 doesn't seem to be running properly
(running `SMTTest`) because `java.library.path` does not seem to be set
correctly.

SMT queries look kind of like the following.
```scala
  def checkPredicate(p: Predicate, timeoutMillis: Int = 50): SatResult = {
    val solver = SMTSolver(timeoutMillis)
    val res = solver.predSat(p, obtainModel = true)
    res match {
      case SatResult.SAT(Some(model)) => println(s"$p has a model $model")
      case _ => {}
    }
    solver.close()
    res
  }
```
This example makes a solver just for a single query, but if you're doing
multiple queries it is better to make a single solver and give it all of
the queries, for performance.

Queries can be given as `Predicate`s, `Expr`s or SMT2 strings like the
ones generated in `IRExpToSMT2.scala`.