discount __bss_start. also include SECTIONS in the normaliseRelf
UNFORTUNATELY, the bss_start correction can't be put in normaliseRelf
alone, because i want to be able to detect when it happens. the mismatch
occurs in the gtsrelf, but its only detectable by considering both the
oldrelf and gtsrelf. if it was in normaliseRelf, the differences would
be smoothed over and undetectable.