remove `Ordered[SpecGlobalOrAccess]` instance
this ordering only considered the address field, which is surprising
when you expect the ordering to impose a total order on the elements.
it is also unlawful because it can claim two values are "equal"
(by being equal in address), even when they are not truly equal in all
of their fields.

i have replaced the old uses of this instance with explicit
`.sortBy(_.address)`.